requirements.txt
bcrypt==4.0.0
click==8.0.0
colorama==0.4.4
Flask==2.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dotenv==0.18.0
Werkzeug==2.0.1
